pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "copybuffer"
version = "1.8.0"
description = "A command-line utility for copying file contents, directory contents, or STDIN input to the system clipboard."
authors = [
    { name = "Draeician", email = "draeician@gmail.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
homepage = "https://github.com/draeician/copybuffer"
repository = "https://github.com/draeician/copybuffer"
keywords = ["clipboard", "utility", "command-line", "python"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyperclip",
    "Pillow",
    "tiktoken"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
]

[tool.pipx]
include = ["copybuffer"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.hatch.build.targets.wheel]
packages = ["copybuffer"]

[project.scripts]
cb = "copybuffer.main:main"

